d7/task-1/main.py
"""
Network Configuration Backup Tool
A Python module for automating configuration backups with Git version control

This tool allows you to:
- Initialize Git repositories for version control
- Backup device configurations using pyATS
- Compare configurations between backups
- Track changes with automated commits
- Log all operations with rotating file handler
"""

import git
import json
import yaml
import datetime
import argparse
import os
import logging
from genie.utils.diff import Diff
from pathlib import Path
from genie.testbed import load as load_testbed
from logging.handlers import TimedRotatingFileHandler
from inventory_tool import read_inventory

# TASK 1: Git repository management

def commit_backup(backup_dir, commit_message):
    """
    Initialize Git repository if needed and commit backup files.
    
    Args:
        backup_dir (str): Directory path containing backup files to commit
        commit_message (str): Descriptive commit message for the backup operation
    """
    try:
        repo = git.Repo(backup_dir)
    except git.exc.InvalidGitRepositoryError:
        repo = git.Repo.init(backup_dir)
    except Exception as e:
        print(f"Unexpected error: {e}")
        return
    
    repo.git.add(A=True)
    repo.index.commit(commit_message)
    print("Backup committed to Git repository.")

# TASK 2: Device backup functionality

def generate_testbed(inventory_file):
    """
    Generate pyATS testbed from inventory CSV file.
    
    Args:
        inventory_file (str): Path to CSV inventory file
        
    Returns:
        dict: Testbed dictionary for pyATS
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

def fetch_running_config(testbed, device_name):
    """
    Retrieve running configuration from specified device.
    
    Args:
        testbed: Loaded pyATS testbed object
        device_name (str): Name of device to backup
        
    Returns:
        str: Configuration text or None if error
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

def backup_config(testbed, backup_dir):
    """
    Backup configurations from all devices in testbed.
    
    Args:
        testbed: Loaded pyATS testbed
        backup_dir (str): Directory to store backup files
        
    Returns:
        list: List of created backup file paths
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

# TASK 3: Configuration comparison functionality

def get_latest_backups(backup_dir, device_name):
    """
    Find the two most recent backup files for a device.
    
    Args:
        backup_dir (str): Directory containing backup files
        device_name (str): Device name to find backups for
        
    Returns:
        tuple: (latest_file, previous_file) or (None, None) if insufficient files
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

def compare_configs(backup_dir, device_name):
    """
    Compare the two most recent configuration backups for a device.
    
    Args:
        backup_dir (str): Directory containing backup files  
        device_name (str): Device name to compare configs for
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

# TASK 4: Logging and CLI functionality

def log_and_print(message, level="info"):
    """
    Log message to file and print to console.
    
    Args:
        message (str): Message to log to file and display to user
        level (str): Log level - "info", "error", or "warning" (default: "info")
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

def main():
    """
    Main function handling CLI arguments and backup orchestration.
    
    This function sets up argument parsing, orchestrates the backup process,
    and handles Git commits and configuration comparison based on CLI options.
    """
    # COMPLETE THE FUNCTION LOGIC HERE
    pass

if __name__ == "__main__":
    # Test the Git repository functionality
    test_backup_dir = "./backups"
    
    # Create the directory and test file
    Path(test_backup_dir).mkdir(exist_ok=True)
    test_file = Path(test_backup_dir) / "test_backup.txt"
    test_file.write_text(f"Test backup created at {datetime.datetime.now()}")
    
    # Test Git operations
    commit_msg = f"Test backup on {datetime.datetime.now():%Y-%m-%d %H:%M:%S}"
    commit_backup(test_backup_dir, commit_msg)