d7/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
arrow==1.3.0
async-lru==2.0.4
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
backports.ssl==0.0.9
backports.ssl_match_hostname==3.7.0.1
bcrypt==4.3.0
bidict==0.23.1
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
ciscoisesdk==2.3.0
cookiecutter==2.6.0
coverage==7.6.12
cryptography==44.0.1
dict2xml==1.7.6
dill==0.3.9
distro==1.9.0
disutils==1.4.32.post2
exceptiongroup==1.2.2
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.1
frozenlist==1.5.0
genie==25.1
genie.libs.clean==25.1
genie.libs.conf==25.1
genie.libs.filetransferutils==25.1
genie.libs.health==25.1
genie.libs.ops==25.1
genie.libs.parser==25.1
genie.libs.robot==25.1
genie.libs.sdk==25.1
genie.telemetry==25.1
genie.trafficgen==25.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
IxNetwork==9.30.2212.8
ixnetwork_restpy==1.6.0
Jinja2==3.1.5
jsonpickle==4.0.2
junit-xml==1.9
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
ncclient==0.6.17
netaddr==1.3.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pluggy==1.5.0
ply==3.11
prettytable==3.15.1
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.4.8
PyYAML==6.0.2
pyats==25.1
pyats.aereport==25.1
pyats.aetest==25.1
pyats.async==25.1
pyats.connections==25.1
pyats.contrib==25.1
pyats.datastructures==25.1
pyats.easypy==25.1
pyats.kleenex==25.1
pyats.log==25.1
pyats.reporter==25.1
pyats.results==25.1
pyats.robot==25.1
pyats.tcl==25.1
pyats.topology==25.1
pyats.utils==25.1
pycparser==2.22
pyftpdlib==2.0.1
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pysmi==1.5.11
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==8.3.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rest.connector==25.1
rich==13.9.4
robotframework==7.2.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
text-unidecode==1.3
tftpy==0.8.0
tomli==2.2.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
unicon==25.1
unicon.plugins==25.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.2
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.35.1
yang.connector==25.1
yarl==1.18.3
